refactor: centralize reference matrix serialization
Backported to the v4.29 catalog model, which still uses top-level reference_blueprints rather than per-target publish_reference metadata.

(cherry picked from commit 3e9c0293668e8b71aa8194ede1e120ac463a2188)
