+++ scripts/blueprint_harness_projects.py
@@ -1,8 +1,10 @@
 from __future__ import annotations
 
+from collections.abc import Callable
 from dataclasses import dataclass, replace
 import json
 from pathlib import Path
+import subprocess
 
 from scripts.blueprint_harness_branches import active_release_branch, normalize_lean_release_ref
 from scripts.blueprint_harness_manifest import (
@@ -488,6 +490,99 @@ def deploy_project_matrix(
                     "toolchain": target.toolchain,
                     "verso_ref": target.verso_ref,
                     "branch": target.branch,
+                    "project_id": project.project_id,
+                    "hash": project.ref,
+                    "artifact_name": deploy_project_artifact_name(project),
+                    "artifact_path": deploy_project_artifact_path(project),
+                }
+            )
+    return {"include": include}
+
+
+def manifest_relative_to_package(manifest_path: Path, package_root: Path) -> Path | None:
+    try:
+        return manifest_path.relative_to(package_root)
+    except ValueError:
+        return None
+
+
+def release_branch_ref_candidates(branch: str) -> tuple[str, ...]:
+    return (f"origin/{branch}", f"refs/remotes/origin/{branch}", branch)
+
+
+def load_branch_project_catalog(
+    *,
+    branch: str,
+    manifest_path: Path,
+    package_root: Path,
+) -> HarnessProjectCatalog:
+    manifest_relpath = manifest_relative_to_package(manifest_path, package_root)
+    if manifest_relpath is None:
+        return load_project_catalog(manifest_path)
+
+    errors: list[str] = []
+    for ref in release_branch_ref_candidates(branch):
+        result = subprocess.run(
+            ["git", "show", f"{ref}:{manifest_relpath.as_posix()}"],
+            cwd=package_root,
+            text=True,
+            capture_output=True,
+            check=False,
+        )
+        if result.returncode == 0:
+            return load_project_catalog_text(result.stdout, f"{ref}:{manifest_relpath.as_posix()}")
+        errors.append(result.stderr.strip() or result.stdout.strip())
+
+    raise ValueError(
+        f"could not load reference project catalog for release branch `{branch}` from `{manifest_relpath}`; "
+        + "; ".join(error for error in errors if error)
+    )
+
+
+def deploy_matrix_from_release_catalogs(
+    controller_catalog: HarnessProjectCatalog,
+    deployable_targets: tuple[HarnessReleaseTarget, ...],
+    catalog_for_target: Callable[[HarnessReleaseTarget], HarnessProjectCatalog],
+) -> dict[str, list[dict[str, object]]]:
+    include: list[dict[str, object]] = []
+    for target in deployable_targets:
+        selected_blueprints = [
+            blueprint
+            for blueprint in controller_catalog.reference_blueprints
+            if blueprint.toolchain == target.toolchain
+        ]
+        selected_project_ids = reference_blueprint_ids_for_toolchain(controller_catalog, target.toolchain)
+        expected_hash_by_project = {blueprint.blueprint: blueprint.hash for blueprint in selected_blueprints}
+        if controller_catalog.reference_blueprints and not selected_blueprints:
+            raise ValueError(
+                f"release target `{target.release_id}` has `deploy_pages: true` but no reference "
+                f"blueprints for toolchain `{target.toolchain}`"
+            )
+        release_catalog = catalog_for_target(target)
+        release_target = release_catalog.release_target(target.release_id)
+        if release_target is None:
+            raise ValueError(
+                f"catalog for branch `{target.branch}` does not define release target `{target.release_id}`"
+            )
+        projects = resolve_projects_for_release(
+            release_catalog,
+            release_target.release_id,
+            selected_project_ids or None,
+        )
+        for project in projects:
+            expected_hash = expected_hash_by_project.get(project.project_id)
+            if expected_hash is not None and project.ref != expected_hash:
+                raise ValueError(
+                    f"reference blueprint `{project.project_id}` for toolchain `{target.toolchain}` resolves "
+                    f"to `{project.ref}` on branch `{target.branch}`, but the deploy catalog declares `{expected_hash}`"
+                )
+            include.append(
+                {
+                    "release_id": release_target.release_id,
+                    "rc": getattr(release_target, "rc", None),
+                    "toolchain": release_target.toolchain,
+                    "verso_ref": release_target.verso_ref,
+                    "branch": release_target.branch,
                     "project_id": project.project_id,
                     "hash": project.ref,
                     "artifact_name": deploy_project_artifact_name(project),
